move requisition case model to requisitions module
