+++ custom/openlmis/commtrack.py
@@ -1,7 +1,6 @@
 import logging
 from django.dispatch import Signal
 from corehq.apps.commtrack.helpers import make_supply_point
-from corehq.apps.commtrack.models import SupplyPointCase, RequisitionCase
 from corehq.apps.hqcase.dbaccessors import \
     get_supply_point_case_in_domain_by_id
 from corehq.apps.programs.models import Program
@@ -14,6 +13,7 @@ from custom.openlmis.api import OpenLMISEndpoint
 from custom.openlmis.exceptions import BadParentException, OpenLMISAPIException
 from corehq.apps.commtrack import const
 from collections import defaultdict
+from custom.requisitions.models import RequisitionCase
 
 requisition_approved = Signal(providing_args=["requisitions"])
 requisition_receipt = Signal(providing_args=["requisitions"])
